build: generate checksums for build files
SHA1, SHA256, and SHA512 checksums are now generated (and uploaded to
releases) for toolchains on all arches. This will allow checksums of
artifacts to be verified as coming from CI. This is particularly useful
during Fomu workshops where the toolchain is distributed on USB drives
to work around local bandwidth limitations.

`set +x` is used to clean up the output by disabling the echo of
commands while checksum calculation is performed and enabled again with
`set -x` once checksums calcs are complete.